refactor: deprecate and remove obsolete UI components
