+++ packages/react/src/box/index.js
@@ -1,9 +1,7 @@
 import Box from './Box';
 import ControlBox from './ControlBox';
-import PseudoBox from './PseudoBox';
 
 export {
   Box,
   ControlBox,
-  PseudoBox, // deprecated
 };
