+++ packages/react/src/deprecated/PseudoBox.js
@@ -1,6 +1,6 @@
 import { useOnce } from '@tonic-ui/react-hooks';
 import React, { forwardRef } from 'react';
-import Box from './Box';
+import { Box } from '../box';
 
 const PseudoBox = forwardRef((props, ref) => {
   useOnce(() => {
