+++ packages/react/src/popper/index.js
@@ -1,7 +1,5 @@
 import Popper from './Popper';
-import PopperArrow from './PopperArrow';
 
 export {
   Popper,
-  PopperArrow,
 };
