+++ packages/react/src/switch/index.js
@@ -1,7 +1,5 @@
 import Switch from './Switch';
-import ToggleSwitch from './ToggleSwitch';
 
 export {
   Switch,
-  ToggleSwitch, // deprecated
 };
