+++ packages/react/src/theme/index.js
@@ -1,9 +1,7 @@
 import ThemeProvider from './ThemeProvider';
 import useTheme from './useTheme';
-import withTheme from './withTheme';
 
 export {
   ThemeProvider,
   useTheme,
-  withTheme,
 };
